// src/api/generated/model/get-all-payroll-runs-status.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */
export type GetAllPayrollRunsStatus = typeof GetAllPayrollRunsStatus[keyof typeof GetAllPayrollRunsStatus];


export const GetAllPayrollRunsStatus = {
  DRAFT: 'DRAFT',
  APPROVED: 'APPROVED',
  PROCESSED: 'PROCESSED',
} as const;