// src/api/generated/model/get-all-payroll-runs-type.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */
export type GetAllPayrollRunsType = typeof GetAllPayrollRunsType[keyof typeof GetAllPayrollRunsType];


export const GetAllPayrollRunsType = {
  REGULAR: 'REGULAR',
  OFF_CYCLE: 'OFF_CYCLE',
  ADJUSTMENT: 'ADJUSTMENT',
} as const;